+++ packages/http-server/src/__tests__/server.oas.spec.ts
@@ -141,33 +141,6 @@ describe.each([['petstore.oas2.json'], ['petstore.oas3.json']])('server %s', fil
     expect(payload).toHaveProperty('name');
   });
 
-  it('returns requested response example using __example property', async () => {
-    const response = await server.fastify.inject({
-      method: 'GET',
-      url: '/pets/123?__example=cat',
-    });
-
-    const payload = JSON.parse(response.payload);
-
-    expect(response.statusCode).toBe(200);
-    expect(payload).toStrictEqual({
-      id: 2,
-      category: {
-        id: 1,
-        name: 'Felis',
-      },
-      tags: [
-        {
-          id: 1,
-          name: 'pet',
-        },
-      ],
-      name: 'Fluffy',
-      status: 'available',
-      photoUrls: [],
-    });
-  });
-
   it('returns 500 with error when a non-existent example is requested', async () => {
     const response = await server.fastify.inject({
       method: 'GET',
@@ -319,8 +292,35 @@ describe.each([['petstore.oas2.json'], ['petstore.oas3.json']])('server %s', fil
       expect(response.statusCode).toBe(200);
     });
 
-    // oas2 does not support overriding servers
+    // oas2 does not support overriding servers and named examples
     if (file === 'petstore.oas3.json') {
+      it('returns requested response example using __example property', async () => {
+        const response = await server.fastify.inject({
+          method: 'GET',
+          url: '/pets/123?__example=cat',
+        });
+
+        const payload = JSON.parse(response.payload);
+
+        expect(response.statusCode).toBe(200);
+        expect(payload).toStrictEqual({
+          id: 2,
+          category: {
+            id: 1,
+            name: 'Felis',
+          },
+          tags: [
+            {
+              id: 1,
+              name: 'pet',
+            },
+          ],
+          name: 'Fluffy',
+          status: 'available',
+          photoUrls: [],
+        });
+      });
+
       describe('and operation overrides global servers', () => {
         it(`when the server is valid then return 200`, async () => {
           const response = await server.fastify.inject({